Allow alternative form of roman numerals
While standard form for Roman numerals never have more than 3 characters of the same kind, It seems that that is just a preference and helps keep the number shorter. There aren't any requirements for the Roman numerals to actually follow such rules and the most common example is when old clocks have "IIII" instead of "IV" for 4.

For more information on this please take a look at https://en.wikipedia.org/wiki/Roman_numerals#Alternative_forms.