Migrate from "is-thirteen" to "systemd-13d"
As the road map for systemd feature creep anticipates transitioning all integers to the the more efficient base-systemd number system, the sooner we get this introduced the better. It will ease users into the transition from using the old inefficient and insecure number systems (binary,hex,decimal etc) to the systemd number system with minimal blowback.