Fixed bug, added German variants.
  Previous comment:
---
  // B just looks like 13 written closer
    "b",
----
b looks nothing like 13, if this is allowed to stay in then the integrity of this package is compromised as this is erroneous behaviour that I will have to catch if I truly want to see if something is 13. I have replaced it with the upper case B, and added the German ß and ẞ because it's 2018 - it's important to be inclusive of other cultures and they kinda look like a 13 if you have bad handwriting.